pyproject.toml
[project]
name = "ojt-time-tracker"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic==1.15.2",
    "bcrypt==4.3.0",
    "blinker==1.9.0",
    "click==8.1.8",
    "colorama==0.4.6",
    "firebase-admin>=6.8.0",
    "flask==3.1.0",
    "flask-bcrypt==1.0.1",
    "flask-login==0.6.3",
    "flask-migrate==4.1.0",
    "flask-sqlalchemy==3.1.1",
    "greenlet==3.1.1",
    "itsdangerous==2.2.0",
    "jinja2==3.1.6",
    "mako==1.3.10",
    "markupsafe==3.0.2",
    "psycopg2==2.9.10",
    "python-dotenv==1.1.0",
    "setuptools>=80.1.0",
    "sqlalchemy==2.0.40",
    "typing-extensions==4.13.2",
    "werkzeug==3.1.3",
]
